requirements.txt
# Core app runtime
openai
python-dotenv
streamlit
requests

# Document ingestion
PyMuPDF

# Retrieval and semantic search
numpy

# Safety / moderation
detoxify

# Optional:
# spaCy-based NER anonymization was removed from the required install because it
# is not needed for the app to run and is currently problematic on Python 3.14.
# If you later want richer anonymization, install spaCy separately on Python 3.11/3.12.
